extract conf helpers
