+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Initial Version, developed for loraham_daemon 108
 - Queue TX packets and use CONF status events.
 - Track CONF event transitions for TX queue tests.
+- Extract bridge CONF event helpers.
