+++ loraham_kiss_tnc/run_tests.sh
@@ -153,6 +153,7 @@ run_test bridge \
   "$script_dir/tcp_server.c" \
   "$script_dir/loraham_sock.c" \
   "$script_dir/bridge.c" \
+  "$script_dir/bridge_conf.c" \
   "$script_dir/ax25.c" \
   "$script_dir/tnc2.c" \
   "$script_dir/kiss.c" \
